requirements.txt
django>=2.1.0,<2.2.0
djangorestframework>=3.9.0,<3.10.0
django-phone-field
psycopg2>=2.7.0,<2.8.0
Pillow>=5.3.0,<5.4.0
opencv-python-headless==4.2.0.32
tensorflow==2.3.1
